adding newcomponents
