+++ client/src/components/profile/leedur/ProfilePageQRScanner/ProfilePageQRScanner.scss
@@ -1,0 +1,81 @@
+@import '../../../../scssStyles/mixins';
+
+.profile-page-qr-code-scanner {
+  margin-top: 20px;
+  color: var(--text-dynamic) !important;
+  align-items: center;
+  p {
+    font-size: 18px;
+    margin-bottom: 10px;
+  }
+}
+
+.profile-page-side-section {
+  max-width: 350px;
+  background-color: var(--container-bg);
+  box-shadow: var(--shadow);
+  border-radius: 25px;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  padding: 15px;
+  @include devices(tablet) {
+    max-width: 90vw;
+    width: 90vw;
+    margin-top: 20px;
+  }
+  color: var(--text-dynamic);
+}
+
+.profile-page-scanned-data {
+  background-color: var(--purple-shades-light);
+  transition: background-color 450ms;
+  padding: 15px 20px;
+  border-radius: 10px;
+  width: 100%;
+  text-align: center;
+  margin-top: 5px;
+}
+
+.profile-page-scanned-data-success {
+  background-color: var(--green-success);
+  animation: pulse 350ms forwards;
+}
+.profile-page-scanned-data-error {
+  background-color: var(--red-error);
+  animation: pulse 350ms forwards;
+}
+
+.profile-page-manual-entry-header {
+  font-size: 25px;
+  margin-top: 25px;
+}
+
+.manual-sign-in-frosh-search-result-container {
+  margin-bottom: 12px;
+  h3 {
+    font-size: 22px;
+  }
+  p {
+    font-size: 17px;
+    margin: 0px;
+  }
+}
+
+.manual-sign-in-frosh-search-result {
+  background-color: var(--purple-shades-light);
+  width: 100%;
+  margin: 5px 0px !important;
+}
+
+@keyframes pulse {
+  0% {
+    transform: scale(1);
+  }
+  50% {
+    transform: scale(1.04);
+  }
+  100% {
+    transform: scale(1);
+  }
+}
